refactor(core): extract core layer with compatibility shims [roadmap:v0.7.4]
Move models, parser->markdown, classification, validate->validation, artifacts, schema, and fs into rac.core. Thin shims remain at the previous top-level paths (rac.parser, rac.validate, rac.classification, rac.artifacts, rac.schema, rac.models, rac.fs) forwarding to rac.core.* so existing imports keep resolving. No behavior change.
