+++ src/rac/core/__init__.py
@@ -1,0 +1,6 @@
+"""RAC core layer.
+
+Domain and specification primitives with no knowledge of the CLI, terminal
+formatting, or Explorer: artifact models, Markdown parsing, classification,
+validation, artifact specs, schema references, and filesystem discovery.
+"""
